fix: 캐싱 리팩토링

+++ PaperMania/Server/Application/UseCase/Auth/LogoutUseCase.cs
@@ -1,4 +1,6 @@
-﻿using Server.Application.Port;
+﻿using Server.Api.Dto.Response;
+using Server.Application.Exceptions;
+using Server.Application.Port;
 using Server.Application.UseCase.Auth.Command;
 
 namespace Server.Application.UseCase.Auth;
@@ -14,6 +16,22 @@ public class LogoutUseCase : ILogoutUseCase
     
     public async Task ExecuteAsync(LogoutCommand request)
     {
+        ValidateInput(request);
+        
+        var userId = await _sessionService.FindUserIdBySessionIdAsync(request.SessionId);
+        if (userId == null)
+            return;
+        
         await _sessionService.DeleteSessionAsync(request.SessionId);
+    }
+    
+    private static void ValidateInput(LogoutCommand request)
+    {
+        if (string.IsNullOrWhiteSpace(request.SessionId))
+        {
+            throw new RequestException(
+                ErrorStatusCode.BadRequest,
+                "INVALID_SESSION_ID");
+        }
     }
 }
